+++ pkg/prom/ha/sharding.go
@@ -4,15 +4,12 @@ import (
 	"context"
 	"hash/fnv"
 	"net/http"
-	"strings"
-	"sync"
 	"time"
 
 	"github.com/cortexproject/cortex/pkg/ring"
 	"github.com/go-kit/kit/log/level"
 	"github.com/golang/protobuf/ptypes/empty"
 	"github.com/grafana/agent/pkg/agentproto"
-	"github.com/grafana/agent/pkg/prom/instance"
 )
 
 // Reshard initiates an entire reshard of the current HA scraping service instance.
@@ -39,20 +36,22 @@ func (s *Server) Reshard(ctx context.Context, _ *agentproto.ReshardRequest) (_ *
 		discoveredConfigs = map[string]struct{}{}
 	)
 
-	configCh, err := s.AllConfigs(ctx)
+	configCh, err := s.store.All(ctx, func(key string) bool {
+		owns, err := s.owns(key)
+		if err != nil {
+			level.Error(s.logger).Log("msg", "failed to detect if key was owned", "key", key, "err", err)
+			return false
+		}
+		return owns
+	})
 	if err != nil {
 		level.Error(s.logger).Log("msg", "failed getting config list when resharding", "err", err)
 		return nil, err
 	}
 	for ch := range configCh {
-		// Applying configs should only fail if the config is invalid
-		err := s.im.ApplyConfig(ch)
-		if err != nil {
-			level.Error(s.logger).Log("msg", "failed to apply config when resharding", "err", err)
-			continue
+		if s.applyConfig(ch.Name, &ch) {
+			discoveredConfigs[ch.Name] = struct{}{}
 		}
-
-		discoveredConfigs[ch.Name] = struct{}{}
 	}
 
 	// Find the set of configs that disappeared from AllConfigs from the last
@@ -74,56 +73,6 @@ func (s *Server) Reshard(ctx context.Context, _ *agentproto.ReshardRequest) (_ *
 	s.configs = discoveredConfigs
 
 	return &empty.Empty{}, nil
-}
-
-// AllConfigs gets all configs known to the KV store.
-func (s *Server) AllConfigs(ctx context.Context) (<-chan instance.Config, error) {
-	keys, err := s.kv.List(ctx, "")
-	if err != nil {
-		return nil, err
-	}
-
-	ch := make(chan instance.Config)
-
-	var wg sync.WaitGroup
-	wg.Add(len(keys))
-	go func() {
-		wg.Wait()
-		close(ch)
-	}()
-
-	for _, key := range keys {
-		go func(key string) {
-			defer wg.Done()
-
-			owns, err := s.owns(key)
-			if err != nil {
-				level.Error(s.logger).Log("msg", "failed to detect if key was owned", "key", key, "err", err)
-				return
-			} else if !owns {
-				// Unowned key, ignore it.
-				return
-			}
-
-			// TODO(rfratto): retries might be useful here
-			v, err := s.kv.Get(ctx, key)
-			if err != nil {
-				level.Error(s.logger).Log("msg", "failed to get config with key", "key", key, "err", err)
-				return
-			} else if v == nil {
-				level.Warn(s.logger).Log("skipping key that was deleted after list was called", "key", key)
-				return
-			}
-
-			cfg, err := instance.UnmarshalConfig(strings.NewReader(v.(string)))
-			if err != nil {
-				level.Error(s.logger).Log("msg", "could not unmarshal stored config", "name", key, "err", err)
-			}
-
-			ch <- *cfg
-		}(key)
-	}
-	return ch, nil
 }
 
 // owns checks to see if a config name is owned by this Server. owns will