+++ pkg/prom/ha/sharding_test.go
@@ -3,12 +3,14 @@ package ha
 import (
 	"context"
 	"sort"
+	"sync"
 	"testing"
 
 	"github.com/cortexproject/cortex/pkg/ring"
-	"github.com/cortexproject/cortex/pkg/ring/kv/consul"
 	"github.com/go-kit/kit/log"
 	"github.com/grafana/agent/pkg/agentproto"
+	"github.com/grafana/agent/pkg/prom/instance"
+	"github.com/grafana/agent/pkg/prom/instance/configstore"
 	"github.com/stretchr/testify/require"
 )
 
@@ -18,13 +20,7 @@ func TestServer_Reshard(t *testing.T) {
 	//	- All configs not in the store but in the existing InstanceManager should be deleted
 	fakeIm := newFakeInstanceManager()
 
-	mockKv := consul.NewInMemoryClient(GetCodec())
-	for _, name := range []string{"keep_a", "keep_b", "new_a", "new_b"} {
-		err := mockKv.CAS(context.Background(), name, func(in interface{}) (out interface{}, retry bool, err error) {
-			return testConfig(t, name), true, nil
-		})
-		require.NoError(t, err)
-	}
+	mockStore := storeWithKeys("keep_a", "keep_b", "new_a", "new_b")
 
 	fakeRing := mockFuncReadRing{
 		GetFunc: func(key uint32, op ring.Operation, bufDescs []ring.InstanceDesc, bufHosts, bufZones []string) (ring.ReplicationSet, error) {
@@ -35,8 +31,8 @@ func TestServer_Reshard(t *testing.T) {
 	srv := Server{
 		logger: log.NewNopLogger(),
 
-		kv: mockKv,
-		im: fakeIm,
+		store: mockStore,
+		im:    fakeIm,
 
 		ring: &fakeRing,
 		addr: "test-server",
@@ -69,13 +65,7 @@ func TestServer_Ownership(t *testing.T) {
 	//	- All configs not in the store but in the existing InstanceManager should be deleted
 	fakeIm := newFakeInstanceManager()
 
-	mockKv := consul.NewInMemoryClient(GetCodec())
-	for _, name := range []string{"owned", "unowned"} {
-		err := mockKv.CAS(context.Background(), name, func(in interface{}) (out interface{}, retry bool, err error) {
-			return testConfig(t, name), true, nil
-		})
-		require.NoError(t, err)
-	}
+	mockStore := storeWithKeys("owned", "unowned")
 
 	var (
 		ownedHash = keyHash("owned")
@@ -95,8 +85,8 @@ func TestServer_Ownership(t *testing.T) {
 	srv := Server{
 		logger: log.NewNopLogger(),
 
-		kv: mockKv,
-		im: fakeIm,
+		store: mockStore,
+		im:    fakeIm,
 
 		ring: &fakeRing,
 		addr: "test-server",
@@ -112,4 +102,66 @@ func TestServer_Ownership(t *testing.T) {
 	}
 	sort.Strings(actual)
 	require.Equal(t, expect, actual)
+}
+
+func storeWithKeys(keys ...string) *configstore.Mock {
+	return &configstore.Mock{
+		AllFunc: func(ctx context.Context, keep func(key string) bool) (<-chan instance.Config, error) {
+			ch := make(chan instance.Config)
+
+			go func() {
+				for _, key := range keys {
+					if keep(key) {
+						ch <- *testConfig(key)
+					}
+				}
+				close(ch)
+			}()
+
+			return ch, nil
+		},
+
+		WatchFunc: func() <-chan configstore.WatchEvent {
+			// Return a watcher that will never emit any events.
+			return make(<-chan configstore.WatchEvent)
+		},
+	}
+}
+
+func newFakeInstanceManager() instance.Manager {
+	var mut sync.Mutex
+	var cfgs = make(map[string]instance.Config)
+
+	return &instance.MockManager{
+		// ListInstances isn't used in this package, so we won't bother to try to
+		// fake it here.
+		ListInstancesFunc: func() map[string]instance.ManagedInstance { return nil },
+
+		ListConfigsFunc: func() map[string]instance.Config {
+			mut.Lock()
+			defer mut.Unlock()
+
+			cp := make(map[string]instance.Config, len(cfgs))
+			for k, v := range cfgs {
+				cp[k] = v
+			}
+			return cp
+		},
+
+		ApplyConfigFunc: func(c instance.Config) error {
+			mut.Lock()
+			defer mut.Unlock()
+			cfgs[c.Name] = c
+			return nil
+		},
+
+		DeleteConfigFunc: func(name string) error {
+			mut.Lock()
+			defer mut.Unlock()
+			delete(cfgs, name)
+			return nil
+		},
+
+		StopFunc: func() {},
+	}
 }