+++ CHANGELOG.md
@@ -1,12 +1,13 @@
 NOTE: FreeBSD builds have not been available since v0.6.0 due to a
-cross-compilation issue. The issue for tracking adding support back
-can be found at [#317](https://github.com/grafana/agent/issues/317).
+cross-compilation issue, but will return in v0.13.0.
 
 # Main (unreleased)
 
 - [FEATURE] Official M1 Mac release builds will now be generated! Look for
   `agent-darwin-arm64` and `agentctl-darwin-arm64` in the release assets.
   (@rfratto)
+
+- [FEATURE] Add support for running as a Windows service (@mattdurham)
 
 - [ENHANCEMENT] Support compression for trace export. (@mdisibio)
 
@@ -20,10 +21,20 @@ can be found at [#317](https://github.com/grafana/agent/issues/317).
 - [BUGFIX] Remote write endpoints that never function across the lifetime of the
   Agent will no longer prevent the WAL from being truncated. (@rfratto)
 
+- [BUGFIX] Bring back FreeBSD support. (@rfratto)
+
+- [CHANGE] The Grafana Cloud Agent has been renamed to the Grafana Agent.
+    (@rfratto)
+
 - [CHANGE] Instance configs uploaded to the Config Store API will no longer be
   stored along with the global Prometheus defaults. This is done to allow
   globals to be updated and re-apply the new global defaults to the configs from
   the Config Store. (@rfratto)
+
+- [CHANGE] The User-Agent header sent for logs will now be
+  `GrafanaAgent/<version>` (@rfratto)
+  
+- [ENHANCEMENT] Upgrade `go.opentelemetry.io/collector` to v0.21.0 (@mapno)
 
 # v0.13.0 (2021-02-25)
 
@@ -303,7 +314,7 @@ files to the new format.
 
 # v0.6.0 (2020-09-04)
 
-- [FEATURE] The Grafana Cloud Agent can now collect logs and send to Loki. This
+- [FEATURE] The Grafana Agent can now collect logs and send to Loki. This
   is done by embedding Promtail, the official Loki log collection client.
   (@rfratto)